Update BuildSystem.cs
I'm scrapping using physics and check spheres and triggers and colliders, it's just not working. Instead going to do what I was originally going to do with parenting all build objects to one main object to be able to check if there are any objects in the area we want to place.