11/12/24 work
Didn't know Array.Length and List.Count existed... 🤦. Still need to go through this code and change a couple of things to stop getting the error that "Index out of bounds of Array" or whatever the error is. Going to go through and rework my logic for less nesting, hopefully that'll fix it but idk.