Update TreeGeneration.cs
Cleaned up the script a little bit, added comments and sorted the variables we're using. Also added some code to save the Texture2D we're creating to an PNG file for easier viewing, but I have that code commented out right now because the code for creating the Texture 2D is definitely broken, and so is the saving system I have most likely.